refactor(getLastActivity_labelèstatus): into CourseController
